arregle errores en el repository y también puse los metodos crud en el controller, falta el metodo GET

+++ src/main/java/microservice/cloud/discount/coupon/infrastructure/persistence/CouponRepositoryJdbcAdapter.java
@@ -1,7 +1,13 @@
 package microservice.cloud.discount.coupon.infrastructure.persistence;
 
+import java.util.List;
+import java.util.Optional;
 import java.util.function.Consumer;
 
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,22 +24,33 @@ import microservice.cloud.discount.shared.domain.value_objects.Id;
 public class CouponRepositoryJdbcAdapter implements CouponRepository{
 
     private final CouponJdbcRepository couponJdbcRepository;
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     @Override
     @Transactional
     public Coupon pessimisticUpdate(Id id, Consumer<Coupon> function) {
-        Coupon coupon = findByIdForUpdate(id);
+        Coupon coupon = findByIdForUpdate(id.value());
         function.accept(coupon);
         couponJdbcRepository.save(toMap(coupon));
 
         return coupon;
     }
 
-    private Coupon findByIdForUpdate(Id id) {
-        CouponEntity entity = couponJdbcRepository.findByIdForUpdate(id.value())
+    private Coupon findByIdForUpdate(String id) {
+        String sql = "SELECT * FROM coupon WHERE id = :id FOR UPDATE";
+        MapSqlParameterSource params = new MapSqlParameterSource("id", id);
+
+        List<CouponEntity> result = namedParameterJdbcTemplate.query(
+            sql, 
+            params, 
+            new BeanPropertyRowMapper<>(CouponEntity.class)
+        );
+
+        CouponEntity coupon = result.stream()
+            .findFirst()
             .orElseThrow(() -> new DataNotFound("Coupon not found"));
 
-        return toMap(entity);
+        return toMap(coupon);
     }
 
     @Transactional
@@ -45,6 +62,12 @@ public class CouponRepositoryJdbcAdapter implements CouponRepository{
     @Transactional
     @Override
     public void delete(Id id) {
+        boolean exists = couponJdbcRepository.existsById(id.value());
+
+        if(!exists) {
+            throw new DataNotFound("Coupon not found");
+        }
+
         couponJdbcRepository.deleteById(id.value());
     }
 
@@ -60,7 +83,6 @@ public class CouponRepositoryJdbcAdapter implements CouponRepository{
     }
 
     private CouponEntity toMap(Coupon coupon) {
-        
         return new CouponEntity(
             coupon.id().value(),
             coupon.discountId().value(),
